+++ src/tokens/PathNegatedToken.ts
@@ -1,0 +1,22 @@
+import { SubPathInNegatedToken } from "./SubPathInNegatedToken";
+import { PathInNegatedToken } from "./PathInNegatedToken";
+import { TokenNode } from "./TokenNode";
+
+
+/**
+ * Token for the negation of a path statement.
+ *
+ * @see {@link https://www.w3.org/TR/sparql11-query/#rPathPrimary}
+ */
+export class PathNegatedToken implements TokenNode {
+	readonly token:"pathNegated" = "pathNegated";
+	readonly path:PathInNegatedToken | SubPathInNegatedToken;
+
+	constructor( path:PathInNegatedToken | SubPathInNegatedToken ) {
+		this.path = path;
+	}
+
+	toString():string {
+		return `!${ this.path }`;
+	}
+}